trips auto generate id, updated deletions of foreign keys, tripController started

+++ src/main/java/com/example/RideShare/controller/TripController.java
@@ -2,18 +2,22 @@ package com.example.RideShare.controller;
 
 import com.example.RideShare.controller.dto.TripDto;
 import com.example.RideShare.controller.exceptions.TripNotFoundException;
+import com.example.RideShare.controller.exceptions.TripWriteUnauthorizedException;
 import com.example.RideShare.controller.exceptions.UserNotFoundException;
 import com.example.RideShare.controller.exceptions.VehicleNotFoundException;
 import com.example.RideShare.model.entity.Trip;
 import com.example.RideShare.model.entity.User;
 import com.example.RideShare.model.entity.Vehicle;
+import com.example.RideShare.model.repository.PassengerRepository;
 import com.example.RideShare.model.repository.TripRepository;
 import com.example.RideShare.model.repository.UserRepository;
 import com.example.RideShare.model.repository.VehicleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @RestController
@@ -24,13 +28,23 @@ public class TripController {
     private final TripRepository repository;
 
     @Autowired
-    private UserRepository userRepository;
-
-    public TripController(TripRepository repository) {this.repository = repository;}
-
+    private final UserRepository userRepository;
 
     @Autowired
-    private VehicleRepository vehicleRepository;
+    private final VehicleRepository vehicleRepository;
+
+    @Autowired
+    private final PassengerRepository passengerRepository;
+
+    public TripController(TripRepository repository,
+                          UserRepository userRepository,
+                          VehicleRepository vehicleRepository,
+                          PassengerRepository passengerRepository) {
+        this.repository = repository;
+        this.userRepository = userRepository;
+        this.vehicleRepository = vehicleRepository;
+        this.passengerRepository = passengerRepository;
+    }
 
     @GetMapping
     public List<Trip> getAllTrips() {
@@ -38,11 +52,11 @@ public class TripController {
     }
 
     @PostMapping
-    public Trip createTrip(@RequestBody TripDto tripDto) {
+    @PostAuthorize("authentication.principal == returnObject.vehicle.owner.email")
+    public Trip createTrip(Authentication authentication, @RequestBody TripDto tripDto) {
         Trip newTrip = new Trip();
-        newTrip.setTripId(tripDto.getTripId());
-        User driver = userRepository.findById(tripDto.getDriverEmail()).orElseThrow(
-                () -> new UserNotFoundException(tripDto.getDriverEmail()));
+        User driver = userRepository.findById(authentication.getName()).orElseThrow(
+                () -> new UserNotFoundException(authentication.getName()));
         newTrip.setDriver(driver);
         Vehicle vehicle = vehicleRepository.findById(tripDto.getLicensePlate()).orElseThrow(
                 () -> new VehicleNotFoundException(tripDto.getLicensePlate()));
@@ -63,9 +77,6 @@ public class TripController {
     public Trip updateTrip(@RequestBody TripDto updatedTripDto, @PathVariable Long tripId) {
         return repository.findById(tripId)
                 .map(trip -> {
-//                    User driver = userRepository.findById(updatedTripDto.getDriverEmail()).orElseThrow(
-//                            () -> new UserNotFoundException(updatedTripDto.getDriverEmail()));
-//                    trip.setDriver(driver);
                     Vehicle vehicle = vehicleRepository.findById(updatedTripDto.getLicensePlate()).orElseThrow(
                             () -> new VehicleNotFoundException(updatedTripDto.getLicensePlate()));
                     trip.setVehicle(vehicle);
@@ -76,17 +87,28 @@ public class TripController {
 
     // Delete Trip
     @DeleteMapping("/{tripId}")
-    public void deleteTrip(@PathVariable Long tripId) {
+    @Transactional
+    public void deleteTrip(Authentication authentication, @PathVariable Long tripId) {
+        Trip tripToDelete = repository.findById(tripId)
+                .orElseThrow(()-> new TripNotFoundException(tripId));
+
+        if (!tripToDelete.getDriver().getEmail().equals(authentication.getName()))
+            throw new TripWriteUnauthorizedException(tripId);
+
+        //delete every passenger instance that had this trip as a foreign key
+        passengerRepository.deleteAllByTripId(tripId);
+
         repository.deleteById(tripId);
     }
 
-    @GetMapping("/searchByName/{searchstring}")
+    @GetMapping("/searchByDriverName/{searchstring}")
     List<Trip> searchByDriver(@PathVariable("searchstring") String searchString){
-        return repository.searchByDriver(searchString);
+        return repository.searchByDriverName(searchString);
     }
+
 
     @GetMapping("/getDriverTrips/{email}")
     List<Trip> getDriverTrips(@PathVariable("email") String email){
-        return repository.getDriverTrips(email);
+        return repository.getByDriverEmail(email);
     }
 }
